+++ src/Console/ListScheduler.php
@@ -6,6 +6,7 @@ namespace Hmazter\LaravelScheduleList\Console;
 use Hmazter\LaravelScheduleList\ScheduleEvent;
 use Hmazter\LaravelScheduleList\ScheduleList;
 use Illuminate\Console\Command;
+use Illuminate\Support\Collection;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -59,7 +60,7 @@ class ListScheduler extends Command
     {
         $events = $this->scheduleList->all();
 
-        if (count($events) === 0) {
+        if ($events->isEmpty()) {
             $this->info('No tasks scheduled');
             return;
         }
@@ -73,9 +74,9 @@ class ListScheduler extends Command
     }
 
     /**
-     * @param array|ScheduleEvent[] $events
+     * @param Collection|ScheduleEvent[] $events
      */
-    protected function outputCronStyle($events)
+    protected function outputCronStyle(Collection $events)
     {
         foreach ($events as $event) {
             $this->line($event->getExpression() . ' ' . $event->getFullCommand());
@@ -83,19 +84,27 @@ class ListScheduler extends Command
     }
 
     /**
-     * @param array|ScheduleEvent[] $events
+     * @param Collection|ScheduleEvent[] $events
      */
     protected function outputTableStyle($events)
     {
         $isVerbosityNormal = $this->output->getVerbosity() === OutputInterface::VERBOSITY_NORMAL;
+        $usesMutexLocks = $events->reduce(function (bool $carry, ScheduleEvent $event) {
+            return $carry || $event->getMutexStatus() !== null;
+        }, false);
+
         $rows = [];
         foreach ($events as $event) {
-            $rows[] = [
+            $row = [
                 'expression' => $event->getExpression(),
                 'next run at' => $event->getNextRunDate()->format('Y-m-d H:i:s'),
                 'command' => $isVerbosityNormal ? $event->getShortCommand() : $event->getFullCommand(),
                 'description' => $event->getDescription(),
             ];
+            if ($usesMutexLocks) {
+                $row['overlapping lock'] = $event->getOverlappingLockDescription();
+            }
+            $rows[] = $row;
         }
 
         $headers = array_keys($rows[0]);